Client: fix reference to temporary
The `this` pointer used for the send function given to
`Event::Event(sendfn)` and `Presence::Presence(sendfn)` was referencing
an object on the stack. With Clang, the object was alive during the
lifetime of the `Event` and `Presence` objects; with GCC, the object is
destroyed during a function return.

Consequently, the bug cannot be observed on the author's OS X computer
but on the Amazon Linux AMI instance.