cmake: allows poco to be found in a non-standard location
I built poco and installed it into /usr/local/poco-1.7.7. This change
allows that version to be found by invoking cmake as:

  $ cmake \
      -DPOCO_INCLUDE_PATH=/usr/local/poco-1.7.7/include \
      -DPOCO_LIBRARY_PATH=/usr/local/poco-1.7.7/lib

The rationale for this change is that the packaged version of poco is
too old on ubuntu xenial. However, for development, I did not want to
install version 1.7.7 in /usr/local, hence the requirement to look in
an alternaive location.

The build-and-test.sh shell fragement could be updated to use this
mechanism too (not done).

Signed-off-by: Andrew McDermott <aim@frobware.com>