ds-client: allow endpoint to be specified on CLI
Allows websocket endpoint to be specified on the command line as
opposed to being hard coded. If nothing is specified it uses the
existing default of "ws://localhost/deepstream".

For example:

    $ ./src/ds-client
    Caught exception "Connection refused"

    $ ./src/ds-client ws://10.175.102.61:6020/deepstream
    successfully logged in to ws://10.175.102.61:6020/deepstream

I also took the liberty to make the code look like and consistently
use common C++ idioms via-a-vis some C++ and some C IO. And I also
chose to reformat it according to the `stroustrup` coding style.

The client now returns and exits with EXIT_SUCCESS when the client can
successfully connect, or EXIT_FAILURE on any other error.

Signed-off-by: Andrew McDermott <aim@frobware.com>