cmake: include(CTest)
This is reqiured to support valgrind.

CTest reads the location of the memory check command (among other
settings) from a file DartConfiguration.tcl in the build directory.
One way to create the dart configuration file is to simply include the
CTest CMake module in your CMakeLists.txt.

The CTest module will find your valgrind installation in /usr/bin and
put a variable MemoryCheckCommand pointing to it in the
DartConfiguration.tcl file.