cmake: don't pollute leaf directories with build artifacts
Put the libaries and binaries that get built into directories at the
top of the build tree rather than in hard-to-find leaf directories.

If you know do:

    $ cd deepstream.io-client-cpp
    $ mkdir ninja-build
    $ cd ninja-build
    $ cmake -G Ninja ..
    $ ninja

You'll end up with the following tree. And now it is easy to find the
binaries that get built.

    $ tree -a bin lib
    bin
    ├── client
    ├── ds-client
    ├── ds-example
    ├── event
    ├── fuzz-me
    ├── impl
    ├── lexer
    ├── make-fuzzer-input
    ├── message
    ├── message_builder
    ├── parser
    ├── presence
    └── random
    lib
    ├── libdeepstream.so
    └── libtest_lexer.a

It's not clear why we have an impl binary but this layout helps to
discover what gets built.

Closes #53

Signed-off-by: Andrew McDermott <aim@frobware.com>