state: simplify transition when already CONNECTED
You can't transition out of OPEN unless you lose the connection. In
the future there will be a CLOSING event, but we don't have that right
now.

Fix ds-example to use presence.get_all(). It was previously commented
out due to 'sudden disconnect' - the sudden disconnects were due to
the bad state transition.

Signed-off-by: Andrew McDermott <aim@frobware.com>