Add Presence subscribe and unsubscribe to ds-example
Verify that presence.subscribe() and presence.unsubscribe() works by
adding callbacks in ds-example.cpp.

With that in place I see:

    bob is online
    bob is offline

printed when a new client is connected. The `offline` doesn't actually
occur because the callback unregisters itself. If you comment that out
you do see both online and offline messages.

Closes #45.

Signed-off-by: Andrew McDermott <aim@frobware.com>