Use flex_target for building lexer components
Rely on the standard CMake flex_target() function to build the lexer
components in lieu of custom commands. Additionally include the
generated lexer.h using double-quotes (i.e., #include "lexer.h") and
ensure that the generated headers are not installed into
INSTALL_DIR/include.

Note: we would like to use DEFINES_FILE from the flex_target() macros
but that is only available in cmake >= 3.5. To workaround this we use
the '%option header-file="lexer.h"' option to specify and generate the
header file.

Closes #84.

Signed-off-by: Andrew McDermott <aim@frobware.com>