Partially revert "Use flex_target for building lexer components"
This partially reverts commit PR #87.

As we had to use older semantics of the flex_target()
function (because of the version of FindFLEX.cmake on trusty) it meant
that Flex's artifacts (lexer.c and lexer.h) polluted the source tree
during the build. By reverting we keep the essence of PR #87 and the
work done in PR #91 ensures that lexer.h does not end up in
INSTALL_DIR/include.

Signed-off-by: Andrew McDermott <aim@frobware.com>