CI: temporarily disable swig on travis trusty
The version of swig in the trusy-backports repo is 3.0.2 which has the
following issue https://github.com/swig/swig/issues/147 (which was fixed
in swig 3.0.3).
The issue leads to name collisions when using `enum class`.
We plan to re-enable swig building on CI later, until then we will
continue to build swig locally with swig 3.0.12.