Fix: use SubscriptionIds to track subscriptions
Previously we used std::unique_ptr<SubscribeFn> so that specific
subscriptions could be unsubscribed (which mimics the javascript API).
However, this has several disadvantages in C++, including space
consumption, heap allocations and indirection.

Now, we instead simply return an integer SubscribeFn that represents the
subscription, and we store the relevant subscribefn in a std::map.